+++ xarray/core/groupers.py
@@ -8,9 +8,8 @@ from __future__ import annotations
 
 import datetime
 from abc import ABC, abstractmethod
-from collections.abc import Mapping, Sequence
 from dataclasses import dataclass, field
-from typing import Any
+from typing import TYPE_CHECKING, Any, Literal, cast
 
 import numpy as np
 import pandas as pd
@@ -21,9 +20,12 @@ from xarray.core.dataarray import DataArray
 from xarray.core.groupby import T_Group, _DummyGroup
 from xarray.core.indexes import safe_cast_to_index
 from xarray.core.resample_cftime import CFTimeGrouper
-from xarray.core.types import DatetimeLike, SideOptions, T_GroupIndices
+from xarray.core.types import Bins, DatetimeLike, GroupIndices, SideOptions
 from xarray.core.utils import emit_user_level_warning
 from xarray.core.variable import Variable
+
+if TYPE_CHECKING:
+    pass
 
 __all__ = [
     "EncodedGroups",
@@ -41,20 +43,25 @@ RESAMPLE_DIM = "__resample_dim__"
 class EncodedGroups:
     """
     Dataclass for storing intermediate values for GroupBy operation.
-    Returned by factorize method on Grouper objects.
+    Returned by the ``factorize`` method on Grouper objects.
 
-    Parameters
+    Attributes
     ----------
-    codes: integer codes for each group
-    full_index: pandas Index for the group coordinate
-    group_indices: optional, List of indices of array elements belonging
-                   to each group. Inferred if not provided.
-    unique_coord: Unique group values present in dataset. Inferred if not provided
+    codes : DataArray
+        Same shape as the DataArray to group by. Values consist of a unique integer code for each group.
+    full_index : pd.Index
+        Pandas Index for the group coordinate containing unique group labels.
+        This can differ from ``unique_coord`` in the case of resampling and binning,
+        where certain groups in the output need not be present in the input.
+    group_indices : tuple of int or slice or list of int, optional
+        List of indices of array elements belonging to each group. Inferred if not provided.
+    unique_coord : Variable, optional
+        Unique group values present in dataset. Inferred if not provided
     """
 
     codes: DataArray
     full_index: pd.Index
-    group_indices: T_GroupIndices | None = field(default=None)
+    group_indices: GroupIndices | None = field(default=None)
     unique_coord: Variable | _DummyGroup | None = field(default=None)
 
     def __post_init__(self):
@@ -69,33 +76,40 @@ class EncodedGroups:
 
 
 class Grouper(ABC):
-    """Base class for Grouper objects that allow specializing GroupBy instructions."""
+    """Abstract base class for Grouper objects that allow specializing GroupBy instructions."""
 
     @property
     def can_squeeze(self) -> bool:
-        """TODO: delete this when the `squeeze` kwarg is deprecated. Only `UniqueGrouper`
-        should override it."""
+        """
+        Do not use.
+
+        .. deprecated:: 2024.01.0
+            This is a deprecated method. It will be deleted when the `squeeze` kwarg is deprecated.
+            Only ``UniqueGrouper`` should override it.
+        """
         return False
 
     @abstractmethod
-    def factorize(self, group) -> EncodedGroups:
+    def factorize(self, group: T_Group) -> EncodedGroups:
         """
-        Takes the group, and creates intermediates necessary for GroupBy.
-        These intermediates are
-        1. codes - Same shape as `group` containing a unique integer code for each group.
-        2. group_indices - Indexes that let us index out the members of each group.
-        3. unique_coord - Unique groups present in the dataset.
-        4. full_index - Unique groups in the output. This differs from `unique_coord` in the
-           case of resampling and binning, where certain groups in the output are not present in
-           the input.
+        Creates intermediates necessary for GroupBy.
 
-        Returns an instance of EncodedGroups.
+        Parameters
+        ----------
+        group : DataArray
+            DataArray we are grouping by.
+
+        Returns
+        -------
+        EncodedGroups
         """
         pass
 
 
 class Resampler(Grouper):
-    """Base class for Grouper objects that allow specializing resampling-type GroupBy instructions.
+    """
+    Abstract base class for Grouper objects that allow specializing resampling-type GroupBy instructions.
+
     Currently only used for TimeResampler, but could be used for SpaceResampler in the future.
     """
 
@@ -106,7 +120,7 @@ class Resampler(Grouper):
 class UniqueGrouper(Grouper):
     """Grouper object for grouping by a categorical variable."""
 
-    _group_as_index: pd.Index | None = None
+    _group_as_index: pd.Index | None = field(default=None, repr=False)
 
     @property
     def is_unique_and_monotonic(self) -> bool:
@@ -117,6 +131,7 @@ class UniqueGrouper(Grouper):
 
     @property
     def group_as_index(self) -> pd.Index:
+        """Caches the group DataArray as a pandas Index."""
         if self._group_as_index is None:
             self._group_as_index = self.group.to_index()
         return self._group_as_index
@@ -127,7 +142,7 @@ class UniqueGrouper(Grouper):
         is_dimension = self.group.dims == (self.group.name,)
         return is_dimension and self.is_unique_and_monotonic
 
-    def factorize(self, group1d) -> EncodedGroups:
+    def factorize(self, group1d: T_Group) -> EncodedGroups:
         self.group = group1d
 
         if self.can_squeeze:
@@ -158,8 +173,9 @@ class UniqueGrouper(Grouper):
         # no need to factorize
         # use slices to do views instead of fancy indexing
         # equivalent to: group_indices = group_indices.reshape(-1, 1)
-        group_indices: T_GroupIndices = [slice(i, i + 1) for i in range(size)]
+        group_indices: GroupIndices = tuple(slice(i, i + 1) for i in range(size))
         size_range = np.arange(size)
+        full_index: pd.Index
         if isinstance(self.group, _DummyGroup):
             codes = self.group.to_dataarray().copy(data=size_range)
             unique_coord = self.group
@@ -179,23 +195,71 @@ class UniqueGrouper(Grouper):
 
 @dataclass
 class BinGrouper(Grouper):
-    """Grouper object for binning numeric data."""
+    """
+    Grouper object for binning numeric data.
 
-    bins: int | Sequence | pd.IntervalIndex
-    cut_kwargs: Mapping = field(default_factory=dict)
-    binned: Any = None
-    name: Any = None
+    Attributes
+    ----------
+    bins : int, sequence of scalars, or IntervalIndex
+        The criteria to bin by.
+
+        * int : Defines the number of equal-width bins in the range of `x`. The
+          range of `x` is extended by .1% on each side to include the minimum
+          and maximum values of `x`.
+        * sequence of scalars : Defines the bin edges allowing for non-uniform
+          width. No extension of the range of `x` is done.
+        * IntervalIndex : Defines the exact bins to be used. Note that
+          IntervalIndex for `bins` must be non-overlapping.
+
+    right : bool, default True
+        Indicates whether `bins` includes the rightmost edge or not. If
+        ``right == True`` (the default), then the `bins` ``[1, 2, 3, 4]``
+        indicate (1,2], (2,3], (3,4]. This argument is ignored when
+        `bins` is an IntervalIndex.
+    labels : array or False, default None
+        Specifies the labels for the returned bins. Must be the same length as
+        the resulting bins. If False, returns only integer indicators of the
+        bins. This affects the type of the output container (see below).
+        This argument is ignored when `bins` is an IntervalIndex. If True,
+        raises an error. When `ordered=False`, labels must be provided.
+    retbins : bool, default False
+        Whether to return the bins or not. Useful when bins is provided
+        as a scalar.
+    precision : int, default 3
+        The precision at which to store and display the bins labels.
+    include_lowest : bool, default False
+        Whether the first interval should be left-inclusive or not.
+    duplicates : {"raise", "drop"}, default: "raise"
+        If bin edges are not unique, raise ValueError or drop non-uniques.
+    """
+
+    bins: Bins
+    # The rest are copied from pandas
+    right: bool = True
+    labels: Any = None
+    precision: int = 3
+    include_lowest: bool = False
+    duplicates: Literal["raise", "drop"] = "raise"
 
     def __post_init__(self) -> None:
         if duck_array_ops.isnull(self.bins).all():
             raise ValueError("All bin edges are NaN.")
 
-    def factorize(self, group) -> EncodedGroups:
+    def factorize(self, group: T_Group) -> EncodedGroups:
         from xarray.core.dataarray import DataArray
 
-        data = group.data
+        data = np.asarray(group.data)  # Cast _DummyGroup data to array
 
-        binned, self.bins = pd.cut(data, self.bins, **self.cut_kwargs, retbins=True)
+        binned, self.bins = pd.cut(  # type: ignore [call-overload]
+            data,
+            bins=self.bins,
+            right=self.right,
+            labels=self.labels,
+            precision=self.precision,
+            include_lowest=self.include_lowest,
+            duplicates=self.duplicates,
+            retbins=True,
+        )
 
         binned_codes = binned.codes
         if (binned_codes == -1).all():
@@ -222,7 +286,51 @@ class BinGrouper(Grouper):
 
 @dataclass
 class TimeResampler(Resampler):
-    """Grouper object specialized to resampling the time coordinate."""
+    """
+    Grouper object specialized to resampling the time coordinate.
+
+    Attributes
+    ----------
+    freq : str
+        Frequency to resample to. See `Pandas frequency
+        aliases <https://pandas.pydata.org/pandas-docs/stable/user_guide/timeseries.html#offset-aliases>`_
+        for a list of possible values.
+    closed : {"left", "right"}, optional
+        Side of each interval to treat as closed.
+    label : {"left", "right"}, optional
+        Side of each interval to use for labeling.
+    base : int, optional
+        For frequencies that evenly subdivide 1 day, the "origin" of the
+        aggregated intervals. For example, for "24H" frequency, base could
+        range from 0 through 23.
+
+        .. deprecated:: 2023.03.0
+            Following pandas, the ``base`` parameter is deprecated in favor
+            of the ``origin`` and ``offset`` parameters, and will be removed
+            in a future version of xarray.
+
+    origin : {"epoch", "start", "start_day", "end", "end_day"}, pandas.Timestamp, datetime.datetime, numpy.datetime64, or cftime.datetime, default: "start_day"
+        The datetime on which to adjust the grouping. The timezone of origin
+        must match the timezone of the index.
+
+        If a datetime is not used, these values are also supported:
+        - 'epoch': `origin` is 1970-01-01
+        - 'start': `origin` is the first value of the timeseries
+        - 'start_day': `origin` is the first day at midnight of the timeseries
+        - 'end': `origin` is the last value of the timeseries
+        - 'end_day': `origin` is the ceiling midnight of the last day
+    offset : pd.Timedelta, datetime.timedelta, or str, default is None
+        An offset timedelta added to the origin.
+    loffset : timedelta or str, optional
+        Offset used to adjust the resampled time labels. Some pandas date
+        offset strings are supported.
+
+        .. deprecated:: 2023.03.0
+            Following pandas, the ``loffset`` parameter is deprecated in favor
+            of using time offset arithmetic, and will be removed in a future
+            version of xarray.
+
+    """
 
     freq: str
     closed: SideOptions | None = field(default=None)
@@ -232,8 +340,8 @@ class TimeResampler(Resampler):
     loffset: datetime.timedelta | str | None = field(default=None)
     base: int | None = field(default=None)
 
-    index_grouper: CFTimeGrouper | pd.Grouper = field(init=False)
-    group_as_index: pd.Index = field(init=False)
+    index_grouper: CFTimeGrouper | pd.Grouper = field(init=False, repr=False)
+    group_as_index: pd.Index = field(init=False, repr=False)
 
     def __post_init__(self):
         if self.loffset is not None:
@@ -275,7 +383,7 @@ class TimeResampler(Resampler):
         if isinstance(group_as_index, CFTimeIndex):
             from xarray.core.resample_cftime import CFTimeGrouper
 
-            index_grouper = CFTimeGrouper(
+            self.index_grouper = CFTimeGrouper(
                 freq=self.freq,
                 closed=self.closed,
                 label=self.label,
@@ -284,7 +392,7 @@ class TimeResampler(Resampler):
                 loffset=self.loffset,
             )
         else:
-            index_grouper = pd.Grouper(
+            self.index_grouper = pd.Grouper(
                 # TODO remove once requiring pandas >= 2.2
                 freq=_new_to_legacy_freq(self.freq),
                 closed=self.closed,
@@ -292,7 +400,6 @@ class TimeResampler(Resampler):
                 origin=self.origin,
                 offset=offset,
             )
-        self.index_grouper = index_grouper
         self.group_as_index = group_as_index
 
     def _get_index_and_items(self) -> tuple[pd.Index, pd.Series, np.ndarray]:
@@ -305,31 +412,34 @@ class TimeResampler(Resampler):
         return full_index, first_items, codes
 
     def first_items(self) -> tuple[pd.Series, np.ndarray]:
-        from xarray import CFTimeIndex
+        from xarray.coding.cftimeindex import CFTimeIndex
+        from xarray.core.resample_cftime import CFTimeGrouper
 
-        if isinstance(self.group_as_index, CFTimeIndex):
-            return self.index_grouper.first_items(self.group_as_index)
-        else:
-            s = pd.Series(np.arange(self.group_as_index.size), self.group_as_index)
-            grouped = s.groupby(self.index_grouper)
-            first_items = grouped.first()
-            counts = grouped.count()
-            # This way we generate codes for the final output index: full_index.
-            # So for _flox_reduce we avoid one reindex and copy by avoiding
-            # _maybe_restore_empty_groups
-            codes = np.repeat(np.arange(len(first_items)), counts)
-            if self.loffset is not None:
-                _apply_loffset(self.loffset, first_items)
-            return first_items, codes
+        if isinstance(self.index_grouper, CFTimeGrouper):
+            return self.index_grouper.first_items(
+                cast(CFTimeIndex, self.group_as_index)
+            )
 
-    def factorize(self, group) -> EncodedGroups:
+        s = pd.Series(np.arange(self.group_as_index.size), self.group_as_index)
+        grouped = s.groupby(self.index_grouper)
+        first_items = grouped.first()
+        counts = grouped.count()
+        # This way we generate codes for the final output index: full_index.
+        # So for _flox_reduce we avoid one reindex and copy by avoiding
+        # _maybe_restore_empty_groups
+        codes = np.repeat(np.arange(len(first_items)), counts)
+        if self.loffset is not None:
+            _apply_loffset(self.loffset, first_items)
+        return first_items, codes
+
+    def factorize(self, group: T_Group) -> EncodedGroups:
         self._init_properties(group)
         full_index, first_items, codes_ = self._get_index_and_items()
         sbins = first_items.values.astype(np.int64)
-        group_indices: T_GroupIndices = [
-            slice(i, j) for i, j in zip(sbins[:-1], sbins[1:])
-        ]
-        group_indices += [slice(sbins[-1], None)]
+        group_indices: GroupIndices = tuple(
+            [slice(i, j) for i, j in zip(sbins[:-1], sbins[1:])]
+            + [slice(sbins[-1], None)]
+        )
 
         unique_coord = Variable(
             dims=group.name, data=first_items.index, attrs=group.attrs
@@ -369,7 +479,7 @@ def _apply_loffset(
         )
 
     if isinstance(loffset, str):
-        loffset = pd.tseries.frequencies.to_offset(loffset)
+        loffset = pd.tseries.frequencies.to_offset(loffset)  # type: ignore[assignment]
 
     needs_offset = (
         isinstance(loffset, (pd.DateOffset, datetime.timedelta))