+++ xarray/tests/test_groupby.py
@@ -13,6 +13,7 @@ from packaging.version import Version
 import xarray as xr
 from xarray import DataArray, Dataset, Variable
 from xarray.core.groupby import _consolidate_slices
+from xarray.core.groupers import BinGrouper, EncodedGroups, Grouper, UniqueGrouper
 from xarray.core.types import InterpOptions
 from xarray.tests import (
     InaccessibleArray,
@@ -113,8 +114,9 @@ def test_multi_index_groupby_map(dataset) -> None:
     assert_equal(expected, actual)
 
 
-def test_reduce_numeric_only(dataset) -> None:
-    gb = dataset.groupby("x", squeeze=False)
+@pytest.mark.parametrize("grouper", [dict(group="x"), dict(x=UniqueGrouper())])
+def test_reduce_numeric_only(dataset, grouper: dict) -> None:
+    gb = dataset.groupby(**grouper, squeeze=False)
     with xr.set_options(use_flox=False):
         expected = gb.sum()
     with xr.set_options(use_flox=True):
@@ -734,7 +736,7 @@ def test_groupby_bins_timeseries() -> None:
     expected = xr.DataArray(
         96 * np.ones((14,)),
         dims=["time_bins"],
-        coords={"time_bins": pd.cut(time_bins, time_bins).categories},
+        coords={"time_bins": pd.cut(time_bins, time_bins).categories},  # type: ignore[arg-type]
     ).to_dataset(name="val")
     assert_identical(actual, expected)
 
@@ -868,10 +870,17 @@ def test_groupby_dataset_errors() -> None:
     with pytest.raises(ValueError, match=r"length does not match"):
         data.groupby(data["dim1"][:3])
     with pytest.raises(TypeError, match=r"`group` must be"):
-        data.groupby(data.coords["dim1"].to_index())
+        data.groupby(data.coords["dim1"].to_index())  # type: ignore[arg-type]
 
 
-def test_groupby_dataset_reduce() -> None:
+@pytest.mark.parametrize(
+    "by_func",
+    [
+        pytest.param(lambda x: x, id="group-by-string"),
+        pytest.param(lambda x: {x: UniqueGrouper()}, id="group-by-unique-grouper"),
+    ],
+)
+def test_groupby_dataset_reduce_ellipsis(by_func) -> None:
     data = Dataset(
         {
             "xy": (["x", "y"], np.random.randn(3, 4)),
@@ -883,10 +892,11 @@ def test_groupby_dataset_reduce() -> None:
 
     expected = data.mean("y")
     expected["yonly"] = expected["yonly"].variable.set_dims({"x": 3})
-    actual = data.groupby("x").mean(...)
+    gb = data.groupby(by_func("x"))
+    actual = gb.mean(...)
     assert_allclose(expected, actual)
 
-    actual = data.groupby("x").mean("y")
+    actual = gb.mean("y")
     assert_allclose(expected, actual)
 
     letters = data["letters"]
@@ -897,7 +907,8 @@ def test_groupby_dataset_reduce() -> None:
             "yonly": data["yonly"].groupby(letters).mean(),
         }
     )
-    actual = data.groupby("letters").mean(...)
+    gb = data.groupby(by_func("letters"))
+    actual = gb.mean(...)
     assert_allclose(expected, actual)
 
 
@@ -1028,15 +1039,29 @@ def test_groupby_bins_cut_kwargs(use_flox: bool) -> None:
     )
     assert_identical(expected, actual)
 
+    with xr.set_options(use_flox=use_flox):
+        actual = da.groupby(
+            x=BinGrouper(bins=x_bins, include_lowest=True, right=False),
+        ).mean()
+    assert_identical(expected, actual)
+
 
 @pytest.mark.parametrize("indexed_coord", [True, False])
-def test_groupby_bins_math(indexed_coord) -> None:
+@pytest.mark.parametrize(
+    ["groupby_method", "args"],
+    (
+        ("groupby_bins", ("x", np.arange(0, 8, 3))),
+        ("groupby", ({"x": BinGrouper(bins=np.arange(0, 8, 3))},)),
+    ),
+)
+def test_groupby_bins_math(groupby_method, args, indexed_coord) -> None:
     N = 7
     da = DataArray(np.random.random((N, N)), dims=("x", "y"))
     if indexed_coord:
         da["x"] = np.arange(N)
         da["y"] = np.arange(N)
-    g = da.groupby_bins("x", np.arange(0, N + 1, 3))
+
+    g = getattr(da, groupby_method)(*args)
     mean = g.mean()
     expected = da.isel(x=slice(1, None)) - mean.isel(x_bins=("x", [0, 0, 0, 1, 1, 1]))
     actual = g - mean
@@ -1624,7 +1649,7 @@ class TestDataArrayGroupBy:
         bins = [0, 1.5, 5]
 
         df = array.to_dataframe()
-        df["dim_0_bins"] = pd.cut(array["dim_0"], bins, **cut_kwargs)
+        df["dim_0_bins"] = pd.cut(array["dim_0"], bins, **cut_kwargs)  # type: ignore[call-overload]
 
         expected_df = df.groupby("dim_0_bins", observed=True).sum()
         # TODO: can't convert df with IntervalIndex to Xarray
@@ -1690,7 +1715,7 @@ class TestDataArrayGroupBy:
         array = DataArray(np.arange(4), [("x", range(4))])
         # one of these bins will be empty
         bins = [0, 4, 5]
-        bin_coords = pd.cut(array["x"], bins).categories
+        bin_coords = pd.cut(array["x"], bins).categories  # type: ignore[call-overload]
         actual = array.groupby_bins("x", bins).sum()
         expected = DataArray([6, np.nan], dims="x_bins", coords={"x_bins": bin_coords})
         assert_identical(expected, actual)
@@ -1701,7 +1726,7 @@ class TestDataArrayGroupBy:
     def test_groupby_bins_multidim(self) -> None:
         array = self.make_groupby_multidim_example_array()
         bins = [0, 15, 20]
-        bin_coords = pd.cut(array["lat"].values.flat, bins).categories
+        bin_coords = pd.cut(array["lat"].values.flat, bins).categories  # type: ignore[call-overload]
         expected = DataArray([16, 40], dims="lat_bins", coords={"lat_bins": bin_coords})
         actual = array.groupby_bins("lat", bins).map(lambda x: x.sum())
         assert_identical(expected, actual)
@@ -2606,3 +2631,45 @@ def test_default_flox_method() -> None:
         assert kwargs["method"] == "cohorts"
     else:
         assert "method" not in kwargs
+
+
+def test_custom_grouper() -> None:
+    class YearGrouper(Grouper):
+        """
+        An example re-implementation of ``.groupby("time.year")``.
+        """
+
+        def factorize(self, group) -> EncodedGroups:
+            assert np.issubdtype(group.dtype, np.datetime64)
+            year = group.dt.year.data
+            codes_, uniques = pd.factorize(year)
+            codes = group.copy(data=codes_).rename("year")
+            return EncodedGroups(codes=codes, full_index=pd.Index(uniques))
+
+    da = xr.DataArray(
+        dims="time",
+        data=np.arange(20),
+        coords={"time": ("time", pd.date_range("2000-01-01", freq="3MS", periods=20))},
+        name="foo",
+    )
+    ds = da.to_dataset()
+
+    expected = ds.groupby("time.year").mean()
+    actual = ds.groupby(time=YearGrouper()).mean()
+    assert_identical(expected, actual)
+
+    actual = ds.groupby({"time": YearGrouper()}).mean()
+    assert_identical(expected, actual)
+
+    expected = ds.foo.groupby("time.year").mean()
+    actual = ds.foo.groupby(time=YearGrouper()).mean()
+    assert_identical(expected, actual)
+
+    actual = ds.foo.groupby({"time": YearGrouper()}).mean()
+    assert_identical(expected, actual)
+
+    for obj in [ds, ds.foo]:
+        with pytest.raises(ValueError):
+            obj.groupby("time.year", time=YearGrouper())
+        with pytest.raises(ValueError):
+            obj.groupby()